requirements.txt
Django==4.0.7 # https://github.com/celery/django-celery-beat/issues/566
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-cleanup==6.0.0
django-templated-mail==1.1.1
django-cloudinary-storage==0.3.0
django-celery-email==3.0.0
django-celery-beat==2.3.0
django-pickling==1.0
django-redis==5.2.0
django-cacheops==6.1
django-money==3.0.0
django-svg-image-form-field==1.0.1

djangorestframework==3.13.1
drf-spectacular==0.23.1
drf_base64==2.0

orjson==3.7.8
psycopg2-binary==2.9.3
redis==4.3.4
Pillow==9.2.0
gunicorn==20.1.0
whitenoise==6.2.0
celery==5.2.7
gevent==21.12.0
argon2-cffi==21.3.0
pydantic[email]==1.9.2
pysha3==1.0.2
ethereum==2.3.2
sentry-sdk==1.7.2
py-moneyed==2.0

# dev
django-silk==5.0.1

# test
parameterized==0.8.1
factory_boy==3.2.1
Faker==14.1
